simple loop order pipeline
