+++ src/finchlite/autoschedule/loop_ordering.py
@@ -1,0 +1,206 @@
+from functools import reduce
+from itertools import chain as join_chains
+from typing import overload
+
+from finchlite.algebra import ffuncs
+from finchlite.algebra.tensor import TensorFType
+from finchlite.algebra.utils import intersect
+from finchlite.finch_logic import (
+    Aggregate,
+    Alias,
+    Field,
+    Literal,
+    LogicExpression,
+    LogicLoader,
+    LogicNode,
+    LogicStatement,
+    MockLogicLoader,
+    Plan,
+    Produces,
+    Query,
+    Relabel,
+    Reorder,
+    StatsFactory,
+    Table,
+    TensorStats,
+)
+from finchlite.symbolic import PostOrderDFS, PostWalk, Rewrite, gensym
+
+from .stages import LogicLoopOrderOptimizer
+from .standardize import concordize, flatten_plans, push_fields
+
+
+def add_aggregates(
+    root: LogicStatement, bindings: dict[Alias, TensorFType]
+) -> LogicStatement:
+    fill_values = root.infer_fill_value(
+        {var: val.fill_value for var, val in bindings.items()}
+    )
+
+    def rule_0(node):
+        match node:
+            case Query(lhs, Reorder(Aggregate(_, _, arg, idxs), _)):
+                return node
+            case Query(lhs, Reorder(arg, idxs)):
+                return Query(
+                    lhs,
+                    Reorder(
+                        Aggregate(
+                            Literal(ffuncs.overwrite),
+                            Literal(fill_values[lhs]),
+                            arg,
+                            (),
+                        ),
+                        idxs,
+                    ),
+                )
+            case Query(lhs, Aggregate(_, _, arg, idxs)):
+                return node
+            case Query(lhs, arg):
+                return Query(
+                    lhs,
+                    Aggregate(
+                        Literal(ffuncs.overwrite), Literal(fill_values[lhs]), arg, ()
+                    ),
+                )
+
+    return Rewrite(PostWalk(rule_0))(root)
+
+class CycleInFields(Exception): ...
+
+
+def toposort(chains: list[list[Field]]) -> tuple[Field, ...]:
+    chains = [c for c in chains if len(c) > 0]
+    parents = {chain[0]: 0 for chain in chains}
+    for chain in chains:
+        for f in chain[1:]:
+            parents[f] = parents.get(f, 0) + 1
+    roots = [f for f in parents if parents[f] == 0]
+    perm = []
+    while len(parents) > 0:
+        if len(roots) == 0:
+            raise CycleInFields("Cycle detected in fields' orders")
+        perm.append(roots.pop())
+        for chain in chains:
+            if len(chain) > 0 and chain[0] == perm[-1]:
+                chain.pop(0)
+                if len(chain) > 0:
+                    parents[chain[0]] -= 1
+                    if parents[chain[0]] == 0:
+                        roots.append(chain[0])
+        parents.pop(perm[-1])
+    return tuple(perm)
+
+
+def _heuristic_loop_order(root: LogicExpression) -> tuple[Field, ...]:
+    chains = []
+    for node in PostOrderDFS(root):
+        match node:
+            case Reorder(Table(_, idxs_1), idxs_2):
+                chains.append(list(intersect(intersect(idxs_1, idxs_2), root.fields())))
+    chains.extend([f] for f in root.fields())
+
+    need_fix = False
+    try:
+        result = toposort(chains)
+    except CycleInFields:
+        import warnings
+
+        warnings.warn("Cycle in fields detected, need to permute.", stacklevel=1)
+        need_fix = True
+        result = root.fields()
+
+    if need_fix or reduce(max, [len(c) for c in chains], 0) < len(
+        set(join_chains(*chains))
+    ):
+        counts: dict[Field, int] = {}
+        for chain in chains:
+            for f in chain:
+                counts[f] = counts.get(f, 0) + 1
+        result = tuple(sorted(result, key=lambda x: counts[x] == 1))
+    return result
+
+
+@overload
+def _set_loop_order(
+    node: LogicStatement, perms: dict[LogicNode, LogicExpression]
+) -> LogicStatement: ...
+@overload
+def _set_loop_order(
+    node: LogicNode, perms: dict[LogicNode, LogicExpression]
+) -> LogicNode: ...
+def _set_loop_order(node, perms):
+    def rule_0(node):
+        match node:
+            case Table(Alias(_) as tns, idxs) if tns in perms:
+                return Relabel(perms[tns], idxs)
+        return None
+
+    match node:
+        case Plan(bodies):
+            return Plan(tuple(_set_loop_order(body, perms) for body in bodies))
+        case Query(lhs, Aggregate(op, init, arg, idxs) as rhs):
+            rhs = push_fields(Rewrite(PostWalk(rule_0))(rhs))
+            assert isinstance(arg, LogicExpression)
+            idxs_2 = _heuristic_loop_order(arg)
+            rhs_2 = Aggregate(op, init, Reorder(arg, idxs_2), idxs)
+            perms[lhs] = Reorder(Table(lhs, tuple(rhs_2.fields())), tuple(rhs.fields()))
+            return Query(lhs, rhs_2)
+        case Query(lhs, Reorder(Table(Alias(_) as tns, _), idxs)) as q:
+            tns = perms.get(tns, tns)
+            perms[lhs] = Table(lhs, idxs)
+            return q
+        case Query(lhs, rhs):  # assuming rhs is a bunch of mapjoins
+            rhs = push_fields(Rewrite(PostWalk(rule_0))(rhs))
+            assert isinstance(rhs, LogicExpression)
+            idxs = _heuristic_loop_order(rhs)
+            perms[lhs] = Reorder(Table(lhs, idxs), tuple(rhs.fields()))
+            rhs_2 = Reorder(rhs, idxs)
+            return Query(lhs, rhs_2)
+        case Produces(args):
+            renames = {a: Alias(gensym("A")) for a in args if a in perms}
+            bodies = tuple([Query(v, perms[k]) for k, v in renames.items()])
+            args_2 = tuple([renames.get(a, a) for a in args])
+            return Plan(bodies + (Produces(args_2),))
+        case _:
+            raise Exception(f"Invalid node: {node} in set_loop_order")
+
+
+@overload
+def set_loop_order(node: LogicStatement) -> LogicStatement: ...
+@overload
+def set_loop_order(node: LogicNode) -> LogicNode: ...
+def set_loop_order(node):
+    return _set_loop_order(node, {})
+
+
+class DefaultLoopOrderer(LogicLoopOrderOptimizer):
+    def __init__(self, ctx: LogicLoader | None = None):
+        if ctx is None:
+            ctx = MockLogicLoader()
+        self.ctx = ctx
+
+    @classmethod
+    def validate_inputs(
+        cls,
+        term: LogicStatement,
+        bindings: dict[Alias, TensorFType],
+        stats: dict[Alias, TensorStats],
+        stats_factory: StatsFactory,
+    ) -> None:
+        return
+
+    def lower(
+        self,
+        prgm: LogicStatement,
+        bindings: dict[Alias, TensorFType],
+        stats: dict[Alias, TensorStats],
+        stats_factory: StatsFactory,
+    ):
+        prgm = set_loop_order(prgm)
+        prgm = push_fields(prgm)
+        prgm = concordize(prgm, bindings)
+        prgm = flatten_plans(prgm)
+        prgm = add_aggregates(prgm, bindings)
+        prgm = flatten_plans(prgm)
+        return self.ctx(prgm, bindings, stats, stats_factory)
